migration-guide/PromptFlow-to-MAF/phase-4-migrate-ops/4b-deployment/requirements.txt
agent-framework>=1.0.0
agent-framework-azure-ai-search>=1.0.0
fastapi
uvicorn
python-dotenv
azure-monitor-opentelemetry
azure-identity
